clean up some function
